Authentication signup and login functionality

+++ routes/reviews.js
@@ -14,15 +14,19 @@ router.post(
     const { rating, comment } = req.body;
     const listing = await Listing.findById(id);
     const newreview = new Review({ rating, comment });
+    if (req.isAuthenticated() == true) {
+      listing.reviews.push(newreview);
+      await newreview.save();
+      await listing.save();
 
-    listing.reviews.push(newreview);
-    await newreview.save();
-    await listing.save();
-
-    console.log("Review added successfully");
-    req.flash("success", "Review added successfully!");
-    req.flash("success", "New listing created successfully!");
-    res.redirect(`/listings/${id}`);
+      console.log("Review added successfully");
+      req.flash("success", "Review added successfully!");
+      req.flash("success", "New listing created successfully!");
+      res.redirect(`/listings/${id}`);
+    } else {
+      req.flash("error", "You need to be logged in to add a review.");
+      res.redirect("/login");
+    }
   })
 );
 
@@ -31,15 +35,19 @@ router.post(
   "/deleteReview/:reviewId",
   wrapAsync(async (req, res) => {
     const { id, reviewId } = req.params;
+    if (req.isAuthenticated() == true) {
+      await Review.findByIdAndDelete(reviewId);
+      await Listing.findByIdAndUpdate(id, {
+        $pull: { reviews: reviewId },
+      });
 
-    await Review.findByIdAndDelete(reviewId);
-    await Listing.findByIdAndUpdate(id, {
-      $pull: { reviews: reviewId },
-    });
-
-    console.log("Review deleted successfully");
-    req.flash("error", "Review deleted successfully!");
-    res.redirect(`/listings/${id}`);
+      console.log("Review deleted successfully");
+      req.flash("error", "Review deleted successfully!");
+      res.redirect(`/listings/${id}`);
+    } else {
+      req.flash("error", "You need to be logged in to delete a review.");
+      res.redirect("/login");
+    }
   })
 );
